atlan_copilot/requirements.txt
streamlit
langgraph
google-generativeai
qdrant-client
pymongo
beautifulsoup4
requests
python-dotenv
motor
pandas
scikit-learn
langchain
langchain-community
